requirements.txt
streamlit==1.32.2
requests==2.31.0
pandas==2.1.4
folium==0.14.0
streamlit-folium==0.17.0
python-dotenv==1.0.0  # Para gerenciar variáveis de ambiente